Benchmark Result:

Before (using utils)
```
goos: windows
goarch: amd64
pkg: github.com/H0llyW00dzZ/fiber2fa
cpu: AMD Ryzen 9 3900X 12-Core Processor
BenchmarkJSONSonicMiddlewareWithInvalidCookie-24         	  103310	     10266 ns/op	    6065 B/op	      29 allocs/op
BenchmarkJSONSonicWithValid2FA-24                        	   56832	     21915 ns/op	    9777 B/op	      68 allocs/op
BenchmarkJSONSonicWithValidCookie-24                     	   91604	     12338 ns/op	    7562 B/op	      44 allocs/op
BenchmarkJSONStdLibraryMiddlewareWithInvalidCookie-24    	  126547	      9426 ns/op	    6000 B/op	      29 allocs/op
BenchmarkJSONStdLibraryMiddlewareWithValid2FA-24         	   50286	     24796 ns/op	    8244 B/op	      70 allocs/op
BenchmarkJSONStdLibraryWithValidCookie-24                	   58072	     19404 ns/op	    7272 B/op	      49 allocs/op
```
After (without utils)
```
goos: windows
goarch: amd64
pkg: github.com/H0llyW00dzZ/fiber2fa
cpu: AMD Ryzen 9 3900X 12-Core Processor
BenchmarkJSONSonicMiddlewareWithInvalidCookie-24         	  118537	      9420 ns/op	    6060 B/op	      29 allocs/op
BenchmarkJSONSonicWithValid2FA-24                        	   58778	     20506 ns/op	    9237 B/op	      66 allocs/op
BenchmarkJSONSonicWithValidCookie-24                     	   96550	     12558 ns/op	    7371 B/op	      41 allocs/op
BenchmarkJSONStdLibraryMiddlewareWithInvalidCookie-24    	  124382	      9629 ns/op	    6001 B/op	      29 allocs/op
BenchmarkJSONStdLibraryMiddlewareWithValid2FA-24         	   47356	     24286 ns/op	    8189 B/op	      68 allocs/op
BenchmarkJSONStdLibraryWithValidCookie-24                	   58508	     19867 ns/op	    7105 B/op	      46 allocs/op
```